+++ Models/Operacion.cs
@@ -5,11 +5,11 @@ namespace backenddb_c.Models;
 
 public partial class Operacion
 {
-    public decimal CodigoOperacion { get; set; }
+    public int CodigoOperacion { get; set; }
 
     public string NombreOperacion { get; set; } = null!;
 
-    public decimal CodigoCajero { get; set; }
+    public int CodigoCajero { get; set; }
 
     public virtual Cajero CodigoCajeroNavigation { get; set; } = null!;
 }
