+++ Models/Pago.cs
@@ -5,13 +5,13 @@ namespace backenddb_c.Models;
 
 public partial class Pago
 {
-    public decimal NumeroPago { get; set; }
+    public int NumeroPago { get; set; }
 
     public DateTime FechaPago { get; set; }
 
     public decimal MontoPago { get; set; }
 
-    public decimal CodigoPrestamo { get; set; }
+    public int CodigoPrestamo { get; set; }
 
     public virtual Prestamo CodigoPrestamoNavigation { get; set; } = null!;
 }
